+++ api/src/main/java/com/shirj/api/service/IUserService.java
@@ -2,7 +2,18 @@ package com.shirj.api.service;
 
 import com.shirj.api.entity.User;
 
+/**
+ * The interface for {@code User} service.
+ *
+ * @author shirj, wisdom12333@iCloud.com.
+ */
 public interface IUserService {
 
-    User queryByUserId(int userId);
+    /**
+     * <p>Query {@code User} by userId.</p>
+     *
+     * @param userId the user id.
+     * @return the {@code User}.
+     */
+    User queryByUserId(long userId);
 }
