+++ svc/src/main/java/com/shirj/svc/service/impl/UserServiceImpl.java
@@ -14,7 +14,7 @@ public class UserServiceImpl implements IUserService {
     UserDAO userDAO;
 
     @Override
-    public User queryByUserId(int userId) {
+    public User queryByUserId(long userId) {
         return userDAO.queryByUserId(userId);
     }
 }
